Initial Solution Fix
